Add files via upload
This is my current project, it's still a work in progress but the plan is that the bot does real time web to scan the live feed of a CS2 marketplace for skins. It monitors prices and finds skins that are listed under their market value. Once an underpriced skin is found, the bot automatically purchases it.